+++ R/count_occurrences_by_grade.R
@@ -156,7 +156,14 @@ s_count_occurrences_by_grade <- function(df,
 
     if (!is.ordered(grade)) {
       grade_lbl <- obj_label(grade)
-      grade <- formatters::with_label(factor(grade, levels = levels(grade), ordered = TRUE), grade_lbl)
+      lvls <- levels(grade)
+      if (sum(grepl("^\\d+$", lvls)) %in% c(0, length(lvls))) {
+        lvl_ord <- lvls
+      } else {
+        lvls[!grepl("^\\d+$", lvls)] <- min(as.numeric(lvls[grepl("^\\d+$", lvls)])) - 1
+        lvl_ord <- levels(grade)[order(as.numeric(lvls))]
+      }
+      grade <- formatters::with_label(factor(grade, levels = lvl_ord, ordered = TRUE), grade_lbl)
     }
 
     df_max <- stats::aggregate(grade ~ id, FUN = max, drop = FALSE)