+++ NEWS.md
@@ -4,6 +4,9 @@
 * Replaced `synthetic_cdisc_data` with refactored `synthetic_cdisc_dataset` function to speed up dataset loading in tests/examples.
 * Updated all tests to use `rcd_2022_06_27` version of cached data.
 
+### Bug Fixes
+* Fixed bug causing incorrect ordering of numeric grade levels when missing grades are present in `s_count_occurrences_by_grade`.
+
 # tern 0.7.10
 
 ### New Features